+++ src/App.test.tsx
@@ -6,7 +6,7 @@ import { TaskContextProvider } from "./features/tasks/context/TaskContext";
 import { CheckboxContextProvider } from "./features/tasks/context/CheckboxContext";
 import { RewardsProvider } from "./features/rewards/context/RewardsContext";
 import PouchDB from "pouchdb";
-import { ITask } from "./domain/Task";
+import { ITask } from "./features/tasks/domain/Task";
 
 let testDb: PouchDB.Database<ITask> | undefined;
