+++ src/components/rewards/__tests__/GracePeriodToast.test.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { render, screen, waitFor, act } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import '@testing-library/jest-dom';
-import { GracePeriodToast } from '../GracePeriodToast';
+import { GracePeriodToast } from '../../../features/rewards/components/GracePeriodToast';
 import { Task, BaseState } from '../../../features/tasks/domain/Task';
 
 // Mock task helper
