+++ src/features/rewards/services/__tests__/EffectsEngine.test.ts
@@ -1,5 +1,5 @@
-import { effectsEngine } from '../../features/rewards/services/EffectsEngine';
-import { EffectDescriptor } from '../../features/rewards/types/theme';
+import { effectsEngine } from '../EffectsEngine';
+import { EffectDescriptor } from '../../types/theme';
 
 describe('EffectsEngine', () => {
   beforeEach(() => {
