+++ src/features/settings/components/DatabaseDiagnostics.tsx
@@ -2,7 +2,7 @@
 import React, { useState, useEffect } from 'react';
 import { css } from '@emotion/react';
 import PouchDB from 'pouchdb';
-import useTaskHooks from '../../hooks/useTaskHooks';
+import useTaskHooks from '../../tasks/hooks/useTaskHooks';
 import { Task, ROOT_ID } from '../../tasks/domain/Task';
 
 interface DatabaseDiagnosticsProps {
