+++ src/features/tasks/hooks/useTaskHooks.test.tsx
@@ -1,11 +1,11 @@
 import PouchDB from "pouchdb";
 import React, { PropsWithChildren } from "react";
 import useTaskHooks from "./useTaskHooks";
-import { BaseState, ITask, Task, TaskID, ROOT_ID } from "../features/tasks/domain/Task";
+import { BaseState, ITask, Task, TaskID, ROOT_ID } from "../domain/Task";
 import {
   TaskContextProvider,
   TaskContextProviderProps,
-} from "../features/tasks/context/TaskContext";
+} from "../context/TaskContext";
 import { renderHook } from "@testing-library/react";
 
 const createWrapper = ({
